Regions kinda work. There's some weird logic going on with gaps, and I'm going to implement scaling eventually but that will require a refactor of the Region and Layout classes so it'll be for after our interim release if it happens.

+++ SnapPeaApp/SnapPeaApp/Models/Region.cs
@@ -7,10 +7,10 @@ namespace SnapPeaApp
 {
     class Region
     {
-        public float Left { get; set; }
-        public float Top { get; set; }
-        public float Width { get; set; }
-        public float Height { get; set; }
+        public int Left { get; set; }
+        public int Top { get; set; }
+        public int Width { get; set; }
+        public int Height { get; set; }
         public int Color { get; set; }
 
         public bool IsPointIn(Point p)
